remember disabled value of element before startLoading
This should fix #85.
ngDisabled was called before startLoading and stopLoading is executed. But stopLoading always sets the disabled attribute to false, no mather to what value the disabled attribute was set by ngDisabled. That's why I save the disabled value to the directives scope before startLoading sets disabled=true and reset the disabled value to the saved on on stopLoading.