Move the init check inside ngModel.$render()
There is a timing issue with loading chosen before initial options are
made available to <select> in some complex angular contexts (dynamic
compile of template with conditionals). Moving the init() back into
ngModel.$render() ensures we have the 1.8.x behavior, timing-wise.

It's not the cleanest solution, but it's a proven one.